+++ src/tree.js
@@ -697,6 +697,7 @@
         options.onNextCategory = getDefaultFunction( options.onNextCategory, null );
         options.onSetCategory = getDefaultFunction( options.onSetCategory, null );
         options.onRefresh = getDefaultFunction( options.onRefresh, null );
+        options.onDestroy = getDefaultFunction( options.onDestroy, null );
 
         return options;
     }
@@ -1049,6 +1050,71 @@
     this.moveToNextCategory = function( elementId ) {
         if ( isDefinedString( elementId ) && _elements_Data.hasOwnProperty( elementId ) ) {
             moveToNextCategory( _elements_Data[ elementId ].options );
+        }
+
+        return this;
+    };
+
+
+    /*
+     * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+     * Public Functions:  Destroying
+     * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+     */
+
+    /**
+     * destroyAll().
+     * 
+     * Reverts all rendered elements to their original state (without render attributes).
+     * 
+     * @public
+     * @fires       onDestroy
+     * 
+     * @returns     {Object}                                                The Tree.js class instance.
+     */
+    this.destroyAll = function() {
+        for ( var elementId in _elements_Data ) {
+            if ( _elements_Data.hasOwnProperty( elementId ) ) {
+                var bindingOptions = _elements_Data[ elementId ].options;
+
+                bindingOptions.currentView.element.innerHTML = _string.empty;
+                bindingOptions.currentView.element.className = _string.empty;
+
+                _parameter_Document.body.removeChild( bindingOptions.currentView.tooltip );
+
+                fireCustomTrigger( bindingOptions.onDestroy, bindingOptions.currentView.element );
+            }
+        }
+
+        _elements_Data = {};
+
+        return this;
+    };
+
+    /**
+     * destroy().
+     * 
+     * Reverts an element to its original state (without render attributes).
+     * 
+     * @public
+     * @fires       onDestroy
+     * 
+     * @param       {string}    elementId                                   The Tree.js element ID to destroy.
+     * 
+     * @returns     {Object}                                                The Tree.js class instance.
+     */
+    this.destroy = function( elementId ) {
+        if ( isDefinedString( elementId ) && _elements_Data.hasOwnProperty( elementId ) ) {
+            var bindingOptions = _elements_Data[ elementId ].options;
+
+            bindingOptions.currentView.element.innerHTML = _string.empty;
+            bindingOptions.currentView.element.className = _string.empty;
+
+            _parameter_Document.body.removeChild( bindingOptions.currentView.tooltip );
+
+            fireCustomTrigger( bindingOptions.onDestroy, bindingOptions.currentView.element );
+
+            delete _elements_Data[ elementId ];
         }
 
         return this;